requirements.txt
# Core Web Framework
# Note: This project intentionally uses only Flask for simplicity and maintainability
# FastAPI and Redis have been removed to reduce complexity and dependencies
flask>=2.0.0  # Core web application framework
flask-login>=0.5.0  # User session management
flask-jwt-extended  # JWT authentication for API routes
flask-wtf  # Form handling and CSRF protection
flask-cors>=3.0.10  # Cross-origin resource sharing support
flask-sockets>=0.2.1  # WebSocket support for Flask
gevent-websocket>=0.10.1  # WebSocket server implementation

# Security and Integration
python-dotenv>=0.19.0  # Environment configuration loading
email-validator  # Email validation for forms and models
pydantic  # Data validation

# Database
neo4j>=4.4.0  # Graph database driver for Neo4j integration
neomodel  # Object Graph Mapper (OGM) for Neo4j

# System Metrics - Minimal Monitoring
psutil  # Basic system resource monitoring

# Task Management
APScheduler  # Task scheduling

# File Management Note:
# When removing files from the project, use Remove-Item PowerShell command first:
# Example: Remove-Item -Path "path/to/file" -Force

# -----------------------------
# Development Dependencies
# -----------------------------
# Only needed for development/testing
# Install with: pip install -r requirements.txt[dev]

# Testing
pytest  # Testing framework
pytest-cov  # Test coverage reporting

# Code Quality
mypy  # Static type checking
black  # Code formatting
flake8  # Code linting
isort  # Import sorting

# Type Stubs
types-flask  # Type hints for Flask
types-flask-cors  # Type hints for Flask-CORS
types-psutil  # Type hints for psutil
types-neomodel  # Type hints for neomodel

# Additional Dependencies
structlog>=21.1.0  # Structured logging
python-json-logger>=2.0.0  # JSON log formatter

# Note: Version numbers are intentionally omitted to use latest stable versions
# Pin specific versions only when required for stability
# See setup.py for any version constraints